renaming: rename webpack files
